requirements.txt
fastapi
uvicorn
yt-dlp
ollama
python-dotenv
python-multipart
pywhispercpp
